nmr|feat: Add `fmt:all`, `fmt:sh`, and root `clean` to default scripts
Add three new root scripts to the default registry: `fmt:sh` (format shell scripts with shfmt), `fmt:all` (run both `fmt` and `fmt:sh`), and `clean` (delegate to workspace-level clean recursively).

Split `registries.ts` into `default-scripts.ts` (data-only script definitions) and `resolve-scripts.ts` (composition functions). The public API is unchanged.

Fix stale `ci` order assertions in `resolve-scripts.test.ts` and `resolver.test.ts` left over from the prior reorder commit.
