+++ packages/nmr/__tests__/resolver.test.ts
@@ -61,7 +61,7 @@ describe('buildRootRegistry', () => {
     });
 
     expect(registry['demo:catwalk']).toBe('pnpx http-server --port=5189');
-    expect(registry.ci).toEqual(['check:strict', 'build']);
+    expect(registry.ci).toEqual(['build', 'check:strict']);
   });
 });
